MyPersonaAI/MyPersonaAI/requirements.txt
elasticsearch
transformers
datasets
peft
accelerate
bitsandbytes
sentencepiece
scipy
gradio
pyannote.audio 
torchaudio
ffmpeg-python
python-dotenv
